webapp/server/pyproject.toml
[tool.poetry]
name = "flaskproject"
version = "0.1.0"
description = ""
authors = ["graystum <graystum.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^2.0.2"
mysql-connector-python = "^8.0.27"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
